+++ tests/display/assoc_ty.rs
