+++ tests/display/formatting.rs
