+++ tests/display/opaque_ty.rs
