+++ tests/display/self_.rs
