+++ tests/display/struct_.rs
