+++ tests/lib.rs
@@ -3,7 +3,7 @@ mod test_util;
 #[macro_use]
 mod test;
 
-mod logging_db;
+mod display;
 mod lowering;
 
 mod integration;
